+++ web/index.ts
@@ -8,6 +8,8 @@ export * as typographyStyles from './styles/typography';
 
 export * from '@cds/common';
 export * from './Box/Box';
+export * from './Box/VStack';
+export * from './Box/HStack';
 export * from './Text/Text';
 export * from './Offset/Offset';
 export * from './types';